Rename compute -> classical_compute
