+++ tket-qsystem/src/extension.rs
@@ -1,6 +1,6 @@
 //! This module defines the Hugr extensions used by tket-qsystem.
-mod compute;
-pub use compute::wasm;
+mod classical_compute;
+pub use classical_compute::wasm;
 pub mod futures;
 pub mod qsystem;
 pub mod random;
