+++ tket-qsystem/src/extension/classical_compute/wasm.rs
